requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio]
aiofiles
aiohttp
